Add: funcionalidade de não deletar funcionarios ativos

+++ Projeto-BD-Bayard/src/main/java/com/bayard/Projeto_BD_Bayard/repository/CaixaRepositorio.java
@@ -115,6 +115,10 @@ public class CaixaRepositorio {
     }
 
     public void excluirCaixa(String cpf) throws SQLException {
+        if (verificarFuncionarioAtivo(cpf)) {
+            throw new IllegalStateException("Caixa está ativo. Não pode ser excluído.");
+        }
+
         String sqlCaixa = "DELETE FROM Caixa WHERE cpf = ?";
         String sqlFuncionario = "DELETE FROM Funcionarios WHERE cpf = ?";
 
@@ -222,6 +226,23 @@ public class CaixaRepositorio {
         }
 
         return lista;
+    }
+
+    public boolean verificarFuncionarioAtivo(String cpf) throws SQLException {
+        String sql = "SELECT verificaFuncionarioAtivo(?)";
+
+        try (Connection conn = ConexaoBD.conectar();
+             PreparedStatement stmt = conn.prepareStatement(sql)) {
+
+            stmt.setString(1, cpf);
+            ResultSet rs = stmt.executeQuery();
+
+            if (rs.next()) {
+                return rs.getBoolean(1);
+            } else {
+                throw new SQLException("Não foi possível verificar se o funcionário está ativo.");
+            }
+        }
     }
 
 }
