+++ Projeto-BD-Bayard/src/main/java/com/bayard/Projeto_BD_Bayard/repository/EstoquistaRepositorio.java
@@ -131,6 +131,10 @@ public class EstoquistaRepositorio {
     }
 
     public void excluirEstoquista(String cpf) throws SQLException {
+        if (verificarFuncionarioAtivo(cpf)) {
+            throw new IllegalStateException("Estoquista está ativo. Não pode ser excluído.");
+        }
+
         String sqlEstoquista = "DELETE FROM Estoquista WHERE cpf = ?";
         String sqlFuncionario = "DELETE FROM Funcionarios WHERE cpf = ?";
 
@@ -240,6 +244,23 @@ public class EstoquistaRepositorio {
         }
 
         return lista;
+    }
+
+    public boolean verificarFuncionarioAtivo(String cpf) throws SQLException {
+        String sql = "SELECT verificaFuncionarioAtivo(?)";
+
+        try (Connection conn = ConexaoBD.conectar();
+             PreparedStatement stmt = conn.prepareStatement(sql)) {
+
+            stmt.setString(1, cpf);
+            ResultSet rs = stmt.executeQuery();
+
+            if (rs.next()) {
+                return rs.getBoolean(1);
+            } else {
+                throw new SQLException("Não foi possível verificar se o funcionário está ativo.");
+            }
+        }
     }
 
 }
